requirements.txt
pillow
pyTelegramBotAPI
